refactor: 지도/추천 쿼리 리뷰 반영
- BBox 지도 조회: visitCount 상관 서브쿼리를 LEFT JOIN + GROUP BY로 치환 (COUNT(hm.user_id))
- 레벨추천 정렬: 경도항에 COS(RADIANS(lat)) 보정 추가
- MountainMapResponse.visitCount 타입을 Long으로 통일

+++ src/main/java/com/semosan/api/domain/mountain/dto/response/MountainMapResponse.java
@@ -8,7 +8,7 @@ public record MountainMapResponse(
         Double latitude,
         Double longitude,
         Boolean visited,
-        Integer visitCount,
+        Long visitCount,
         String imageUrl
 ) {
 
@@ -20,7 +20,7 @@ public record MountainMapResponse(
                 projection.getLatitude(),
                 projection.getLongitude(),
                 count > 0,
-                (int) count,
+                count,
                 projection.getImageUrl()
         );
     }
